backend/requirements.txt
fastapi
uvicorn>=0.20.0
hypercorn
openai
python-dotenv
requests
pymavlink
python-multipart